+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/competitions/CompetitionIdHelper.java
@@ -1,32 +1,26 @@
 package de.mephisto.vpin.server.competitions;
 
+import com.drew.metadata.StringValue;
 import de.mephisto.vpin.restclient.competitions.CompetitionType;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
+import org.apache.commons.lang3.StringUtils;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class CompetitionIdFactory {
+public class CompetitionIdHelper {
 
   @NonNull
   public static List<CompetitionType> getCompetitionTypes(@Nullable String value) {
     List<CompetitionType> types = new ArrayList<>();
-    if (value != null) {
+    if (!StringUtils.isEmpty(value)) {
       for (CompetitionType competitionType : CompetitionType.values()) {
-        if (value.contains(competitionType.name())) {
+        if (value.contains(competitionType.name().toLowerCase())) {
           types.add(competitionType);
         }
       }
     }
     return types;
-  }
-
-  @NonNull
-  public static String createId(@NonNull Competition competition, boolean isOwner) {
-    if (!isOwner) {
-      return "vps://competition/" + competition.getType().toLowerCase() + "/remote/" + competition.getId();
-    }
-    return "vps://competition/" + competition.getType().toLowerCase() + "/local/" + competition.getId();
   }
 }