+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/competitions/CompetitionIdUpdater.java
@@ -24,9 +24,6 @@ public class CompetitionIdUpdater implements CompetitionChangeListener, Initiali
   private final static Logger LOG = LoggerFactory.getLogger(CompetitionIdUpdater.class);
 
   @Autowired
-  private FrontendStatusService frontendStatusService;
-
-  @Autowired
   private CompetitionService competitionService;
 
   @Autowired