+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/preferences/PreferenceType.java
@@ -1,5 +1,5 @@
 package de.mephisto.vpin.ui.preferences;
 
 public enum PreferenceType {
-  serverSettings, uiSettings, backglassServer, validationSettings
+  serverSettings, uiSettings, backglassServer, validationSettings, competitionSettings
 }