+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/TableFilterController.java
@@ -4,7 +4,9 @@ import de.mephisto.vpin.restclient.PreferenceNames;
 import de.mephisto.vpin.restclient.emulators.GameEmulatorRepresentation;
 import de.mephisto.vpin.restclient.frontend.FrontendType;
 import de.mephisto.vpin.restclient.games.*;
+import de.mephisto.vpin.restclient.iscored.IScoredSettings;
 import de.mephisto.vpin.restclient.playlists.PlaylistRepresentation;
+import de.mephisto.vpin.restclient.preferences.PreferenceChangeListener;
 import de.mephisto.vpin.restclient.preferences.UISettings;
 import de.mephisto.vpin.ui.tables.dialogs.TableDataController;
 import de.mephisto.vpin.ui.tables.models.TableStatus;
@@ -25,8 +27,7 @@ import java.util.function.Predicate;
 
 import static de.mephisto.vpin.ui.Studio.client;
 
-public class TableFilterController extends BaseFilterController<GameRepresentation, GameRepresentationModel>
-    implements Initializable {
+public class TableFilterController extends BaseFilterController<GameRepresentation, GameRepresentationModel> implements Initializable, PreferenceChangeListener {
 
   @FXML
   private CheckBox missingAssetsCheckBox;
@@ -54,6 +55,9 @@ public class TableFilterController extends BaseFilterController<GameRepresentati
 
   @FXML
   private CheckBox noVpsMappingTableCheckBox;
+
+  @FXML
+  private CheckBox iScoredCompetitionCheckBox;
 
   @FXML
   private CheckBox noVpsMappingVersionCheckBox;
@@ -150,6 +154,7 @@ public class TableFilterController extends BaseFilterController<GameRepresentati
       noHighscoreSettingsCheckBox.setSelected(filterSettings.isNoHighscoreSettings());
       noHighscoreSupportCheckBox.setSelected(filterSettings.isNoHighscoreSupport());
       noVpsMappingTableCheckBox.setSelected(filterSettings.isNoVpsTableMapping());
+      iScoredCompetitionCheckBox.setSelected(filterSettings.isIScored());
       noVpsMappingVersionCheckBox.setSelected(filterSettings.isNoVpsVersionMapping());
       withBackglassCheckBox.setSelected(filterSettings.isWithBackglass());
       withPupPackCheckBox.setSelected(filterSettings.isWithPupPack());
@@ -165,6 +170,8 @@ public class TableFilterController extends BaseFilterController<GameRepresentati
 
   @Override
   public void initialize(URL location, ResourceBundle resources) {
+    iScoredCompetitionCheckBox.managedProperty().bindBidirectional(iScoredCompetitionCheckBox.visibleProperty());
+
     configurationFilters.managedProperty().bindBidirectional(configurationFilters.visibleProperty());
     tableAssetFilters.managedProperty().bindBidirectional(tableAssetFilters.visibleProperty());
     vpsFilters.managedProperty().bindBidirectional(vpsFilters.visibleProperty());
@@ -204,6 +211,11 @@ public class TableFilterController extends BaseFilterController<GameRepresentati
     notPlayedCheckBox.setSelected(filterSettings.isNotPlayed());
     notPlayedCheckBox.selectedProperty().addListener((observable, oldValue, newValue) -> {
       filterSettings.setNotPlayed(newValue);
+      applyFilters();
+    });
+    iScoredCompetitionCheckBox.setSelected(filterSettings.isIScored());
+    iScoredCompetitionCheckBox.selectedProperty().addListener((observable, oldValue, newValue) -> {
+      filterSettings.setIScored(newValue);
       applyFilters();
     });
     noHighscoreSettingsCheckBox.setSelected(filterSettings.isNoHighscoreSettings());
@@ -301,6 +313,9 @@ public class TableFilterController extends BaseFilterController<GameRepresentati
       }
       applyFilters();
     });
+
+    client.getPreferenceService().addListener(this);
+    preferencesChanged(PreferenceNames.ISCORED_SETTINGS, null);
   }
 
   public void setEmulator(GameEmulatorRepresentation value) {
@@ -313,4 +328,11 @@ public class TableFilterController extends BaseFilterController<GameRepresentati
     configurationIssuesFilter.setVisible(vpxMode);
   }
 
+  @Override
+  public void preferencesChanged(String key, Object value) {
+    if (PreferenceNames.ISCORED_SETTINGS.equalsIgnoreCase(key)) {
+      IScoredSettings iScoredSettings = client.getPreferenceService().getJsonPreference(PreferenceNames.ISCORED_SETTINGS, IScoredSettings.class);
+      iScoredCompetitionCheckBox.setVisible(iScoredSettings.isEnabled());
+    }
+  }
 }