+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@
   - game multiscore
   - added badge
 - cometition sidebar
+- new error icons
 - **System Manager**: The overall update check is not blocking the UI anymore. This way, you can immediately switch to other system manager tabs now.
 - **Table Data Manager**: The comments dialog has been integrated into the Table Data Manager dialog. The feature was a bit too hidden.
 - **Table Data Manager**: The Table Data Manager dialog has an additional tab "Playlists" now where the game can be assigned to playlists.